refactor: product validate 수정
