+++ src/domain/Product.validateFuncs.ts
@@ -1,10 +1,6 @@
 import { ERR_PRODUCT } from '../constants/errorMessage';
 import { PRODUCT_CONDITION } from '../constants/domain';
 
-export const getRandomNumZeroToMax = (max: number): number => {
-  return Math.floor(Math.random() * (max + 1));
-};
-
 export const validateAllProductProps = (name: string, price: number, quantity: number) => {
   try {
     validateProductName(name);
@@ -19,7 +15,6 @@ export function validateProductName(name: string) {
   if (name.length > PRODUCT_CONDITION.MAX_NAME_LENGTH) {
     throw new Error(ERR_PRODUCT.LONG_NAME);
   }
-  return;
 }
 
 export function validateProductPrice(price: number) {
@@ -29,7 +24,6 @@ export function validateProductPrice(price: number) {
   if (price % PRODUCT_CONDITION.UNIT_PRICE !== 0) {
     throw new Error(ERR_PRODUCT.SMALL_INPUT_THAN_UNIT);
   }
-  return;
 }
 
 export function validateProductQuantity(quantity: number) {
@@ -39,5 +33,4 @@ export function validateProductQuantity(quantity: number) {
   if (quantity < 0) {
     throw new Error(ERR_PRODUCT.NEGATIVE_QUANTITY);
   }
-  return;
 }
