+++ src/utils/utils.ts
@@ -1,0 +1,3 @@
+export const getRandomNumZeroToMax = (max: number): number => {
+  return Math.floor(Math.random() * (max + 1));
+};
